Fix ordering issues, hopefully
In the previous design, the master thread would toggle the notification
flags for all threads if they have not been launched yet. However, the
threads use a tree structure for synchronization, for which they need to
read the old notification flags without explicit synchronization. This
is fine if they wrote that flag themselves, but if the flag was written
by the master thread they may still see the previous value, which leads
to an ABA-type issue (assume for simplicity that the tree has depth 1):

- master thread sets newSense[0:N] to 1 and launches threads
- thread 0 wakes up, does work
- for i in 1:N, thread 0 waits for thread i by reading newSense[i] and
  spin-waiting until sense[i] == newSense[i]; however, newSense[i] was
  written by the master thread, hence thread 0 may see newSense[i] == 0
  and hence consider the waiting condition to be satisfied
- thread 0 signals completion by setting sense[0] to 1
- the master thread waits until sense[0] is 1
- the master thread then stores the next task and sets newSense[0] to 0
- thread 0 waits for newSense[0] to become 0 and then notifies threads
  1:N by toggling newSense[1:N], after which newSense[1:N] is 0
- thread i wakes up and waits indefinitely because newSense[i] equals
  sense[i]

The simple fix is to always use tree-based notification. It would
probably make sense to also execute the forks and joins in the tree.